refactor: extract UserEditCard into its own card component
Moves the inline component out of ChatMessage and into cards/, wrapped in
CardWrapper so it matches the other agent cards.
